pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.46.0",
]
